requirements.txt
bcrypt==3.1.7
boto3==1.14.9
botocore==1.17.9
cffi==1.14.0
cryptography==2.9.2
docutils==0.15.2
jmespath==0.10.0
paramiko==2.7.1
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.1
python-dotenv==0.13.0
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.9
